**MAJOR** LocalUserInfo removed, to LocalUser only

+++ lib/modules/gift_receiver/gift_request_service.dart
@@ -6,6 +6,7 @@ import 'package:alokito_new/models/gift_giver/gift_giver.dart';
 import 'package:alokito_new/models/gift_giver/gift_request.dart';
 import 'package:alokito_new/models/gift_giver/my_position.dart';
 import 'package:alokito_new/models/notification/gift_notification.dart';
+import 'package:alokito_new/shared/config.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:geoflutterfire/geoflutterfire.dart';
@@ -24,13 +25,10 @@ abstract class BaseGiftRequestService {
   Future<GiftReqeust> getGiftRequest({required String id});
 
   Future<void> updateGiftRequestStatus(
-      {required String requesterId,
-      required String giftId,
-      required GiftRequestStatus giftRequestStatus});
+      {required String requesterId, required String giftId, required GiftRequestStatus giftRequestStatus});
 
   Future<bool> changeRequestStatus(
-      {required GiftRequestStatus giftRequestStatus,
-      required GiftNotification giftNotification});
+      {required GiftRequestStatus giftRequestStatus, required GiftNotification giftNotification});
 }
 
 class GiftRequestService implements BaseGiftRequestService {
@@ -40,14 +38,11 @@ class GiftRequestService implements BaseGiftRequestService {
 
   @override
   Future<void> updateGiftRequestStatus(
-      {required String requesterId,
-      required String giftId,
-      required GiftRequestStatus giftRequestStatus}) async {
+      {required String requesterId, required String giftId, required GiftRequestStatus giftRequestStatus}) async {
     await _firestore
         .collection('gift_requests')
         .doc('${requesterId}.$giftId')
-        .update(
-            {'giftRequestStatus': giftRequestStatusToJson(giftRequestStatus)});
+        .update({'giftRequestStatus': giftRequestStatusToJson(giftRequestStatus)});
   }
 
   @override
@@ -62,8 +57,7 @@ class GiftRequestService implements BaseGiftRequestService {
 
   @override
   Future<bool> changeRequestStatus(
-      {required GiftRequestStatus giftRequestStatus,
-      required GiftNotification giftNotification}) async {
+      {required GiftRequestStatus giftRequestStatus, required GiftNotification giftNotification}) async {
     try {
       var docId = '${giftNotification.requesterUid}.${giftNotification.giftId}';
       var doc = await _firestore.collection('gift_requests').doc(docId).get();
@@ -71,14 +65,10 @@ class GiftRequestService implements BaseGiftRequestService {
       if (doc.exists) {
         GiftReqeust giftReqeust = GiftReqeust.fromJson(doc.data() ?? {});
 
-        GiftReqeust giftReqeust1 = giftReqeust.maybeMap(
-            (value) => value.copyWith(giftRequestStatus: giftRequestStatus),
+        GiftReqeust giftReqeust1 = giftReqeust.maybeMap((value) => value.copyWith(giftRequestStatus: giftRequestStatus),
             orElse: () => giftReqeust);
 
-        await _firestore
-            .collection('gift_requests')
-            .doc(docId)
-            .update(giftReqeust1.toJson());
+        await _firestore.collection('gift_requests').doc(docId).update(giftReqeust1.toJson());
       } else {
         print('request doc does not exists');
         return false;
@@ -92,12 +82,9 @@ class GiftRequestService implements BaseGiftRequestService {
   }
 
   @override
-  Future<bool> increaseNoOfTimesGiftRequested(
-      {required GiftGiver giftGiver}) async {
+  Future<bool> increaseNoOfTimesGiftRequested({required GiftGiver giftGiver}) async {
     var currentUserUid = _auth.currentUser?.uid;
-    var docRef = _firestore
-        .collection('gift_requested_data')
-        .doc('${currentUserUid}.${giftGiver.id}');
+    var docRef = _firestore.collection('gift_requested_data').doc('${currentUserUid}.${giftGiver.id}');
 
     try {
       var doc = await docRef.get();
@@ -127,10 +114,7 @@ class GiftRequestService implements BaseGiftRequestService {
     var currentUserUid = _auth.currentUser?.uid;
 
     try {
-      var result = await _firestore
-          .collection('gift_requests')
-          .doc('${currentUserUid}.${giftGiver.id}')
-          .delete();
+      var result = await _firestore.collection('gift_requests').doc('${currentUserUid}.${giftGiver.id}').delete();
 
       print('Service: delete gift request');
 
@@ -146,10 +130,7 @@ class GiftRequestService implements BaseGiftRequestService {
     var currentUserUid = _auth.currentUser?.uid;
 
     try {
-      var result = await _firestore
-          .collection('gift_requests')
-          .doc('${currentUserUid}.${giftGiver.id}')
-          .get();
+      var result = await _firestore.collection('gift_requests').doc('${currentUserUid}.${giftGiver.id}').get();
 
       print('IN finding gift');
       print(result.data());
@@ -172,31 +153,25 @@ class GiftRequestService implements BaseGiftRequestService {
 
     var currentPos = await Location().getLocation();
 
-    var myLocation = geo.point(
-        latitude: currentPos.latitude ?? 23.7590,
-        longitude: currentPos.longitude ?? 90.4119);
+    var myLocation = geo.point(latitude: currentPos.latitude ?? 23.7590, longitude: currentPos.longitude ?? 90.4119);
     var pos = myLocation.data as Map<dynamic, dynamic>;
 
-    MyPosition myPosition = MyPosition(
-        geohash: pos['geohash'] as String,
-        geopoint: pos['geopoint'] as GeoPoint);
+    MyPosition myPosition = MyPosition(geohash: pos['geohash'] as String, geopoint: pos['geopoint'] as GeoPoint);
     var currentUserUid = _auth.currentUser?.uid;
 
     try {
-      var docRef = _firestore
-          .collection('gift_requests')
-          .doc('${currentUserUid}.${giftGiver.id}');
+      var docRef = _firestore.collection('gift_requests').doc('${currentUserUid}.${giftGiver.id}');
 
       GiftReqeust giftReqeust = GiftReqeust(
           id: '${currentUserUid}.${giftGiver.id}',
           giftRequestStatus: GiftRequestStatus.requestPedning,
           giftArea: giftGiver.area,
           giftPosition: giftGiver.position,
-          giftOfferedByRequester: authController.currentUserGiftOffered.value,
-          giftReceivedByRequester: authController.currentUserGiftReceived.value,
-          requesterAvgRating: authController.currentUserRating.value,
-          requesterRatingSum: authController.currentUserRatingSum.value,
-          requesterTotRating: authController.currentUserTotalRating.value,
+          giftOfferedByRequester: authController.currentUser.value.giftOffered,
+          giftReceivedByRequester: authController.currentUser.value.giftReceived,
+          requesterAvgRating: authController.currentUser.value.averageRating,
+          requesterRatingSum: authController.currentUser.value.ratingSum,
+          requesterTotRating: authController.currentUser.value.totalRating,
           giftFor: giftGiver.giftFor,
           giftId: giftGiver.id!,
           giverUid: giftGiver.uid,
@@ -206,9 +181,9 @@ class GiftRequestService implements BaseGiftRequestService {
           giftImageUrl: giftGiver.imageUrl,
           giftDetails: giftGiver.giftDetails,
           requesterPosition: myPosition,
-          requesterName: authController.currentUserName.value,
+          requesterName: authController.currentUser.value.userName,
           giftGiverImageUrl: giftGiver.userImageUrl,
-          requesterImageUrl: authController.currentUserImageUrl.value,
+          requesterImageUrl: authController.currentUser.value.imageUrl ?? PLACEHOLDER_IMAGE_URL,
           createdAt: Timestamp.now());
 
       await docRef.set(giftReqeust.toJson());
